+++ lib/drawing.ex
@@ -34,6 +34,7 @@ defmodule Smith.Drawing do
           OCEx.Shape.t()
           | Smith.Model.t()
           | Smith.Sketch.t()
+          | Smith.Text.t()
           | Smith.Path.t()
           | Assembly.t()
           | Result.t()